+++ src/main/java/org/openlmis/referencedata/dto/UserDto.java
@@ -41,7 +41,6 @@ public final class UserDto extends BaseDto implements User.Exporter, User.Import
   private String timezone;
   private UUID homeFacilityId;
   private boolean active;
-  private boolean loginRestricted;
   private Map<String, Object> extraData;
   private Set<RoleAssignmentDto> roleAssignments = Sets.newHashSet();
 }